refactor: Move performance components to feature-based structure
- Relocate all performance components from src/components/performance/ to src/features/performance/components/
- Update import paths in Dashboard.tsx and Dashboard.test.tsx
- Fix test file import paths for AuthContext
- Align with recommended feature-based project structure from CLAUDE.md

This change improves code organization by:
- Grouping related performance functionality together
- Following the established feature-based architecture pattern
- Making the codebase more maintainable and scalable
