--- src/components/performance/PerformanceDashboard.d.ts
@@ -1,3 +1,0 @@
-/** Performance monitoring dashboard component. */
-import React from 'react';
-export declare const PerformanceDashboard: React.FC;
