--- src/components/performance/PerformancePredictionDashboard.d.ts
@@ -1,6 +1,0 @@
-/**
- * Performance Prediction Dashboard component with AI-powered insights.
- */
-import React from 'react';
-declare const PerformancePredictionDashboard: React.FC;
-export default PerformancePredictionDashboard;
