--- src/components/performance/PerformancePredictionDashboard.test.d.ts
@@ -1,4 +1,0 @@
-/**
- * Test suite for Performance Prediction Dashboard component.
- */
-export {};
