--- src/components/performance/SmartInsightsPanel.d.ts
@@ -1,3 +1,0 @@
-/** Smart Performance Insights Panel Component. */
-import React from 'react';
-export declare const SmartInsightsPanel: React.FC;
