--- src/components/performance/SmartInsightsPanel.test.d.ts
@@ -1,2 +1,0 @@
-/** Tests for Smart Performance Insights Panel Component. */
-export {};
