+++ src/features/performance/components/PerformancePredictionDashboard.test.tsx
@@ -6,7 +6,7 @@ import React from 'react';
 import { render, screen, waitFor, fireEvent, within } from '@testing-library/react';
 import { vi, describe, it, expect, beforeEach, afterEach } from 'vitest';
 import PerformancePredictionDashboard from './PerformancePredictionDashboard';
-import { useAuth } from '../../features/auth/context/AuthContext';
+import { useAuth } from '../../auth/context/AuthContext';
 
 // Mock the dependencies
 vi.mock('../../api/performanceApi', () => ({
@@ -19,7 +19,7 @@ vi.mock('../../api/performanceApi', () => ({
   },
 }));
 
-vi.mock('../../features/auth/context/AuthContext');
+vi.mock('../../auth/context/AuthContext');
 // No chart mocking needed - using CSS-based visualizations
 
 describe('PerformancePredictionDashboard', () => {
