+++ src/features/performance/components/SmartInsightsPanel.test.tsx
