+++ src/features/performance/components/SmartInsightsPanel.tsx
